requirements.txt
python-telegram-bot
cryptocompare
python-dotenv
pytz